==> scripts/01_metadata_extraction.Rmd <==
---
title: "01_metadata_extraction"
author: "Arthur de Grandpré, UQTR @ arthur.de.grandpre@uqtr.ca"
date: "29 jan 2021"
output: 
  html_document: 
    toc: yes
    toc_float: yes
---

# Intro
This is the first script of the Open HRRS W2 workflow for open source high resolution remote sensing of vegetation cover in optically complex waters.  
  
This first script is used to extract and validate metadata from high resolution images. It is adapted to the data structure of DigitalGlobe products, including Quickbird and Worlview-02, 03 & 04.

# 1. Setting up the R environment

## 1.1. Libraries

All libraries must be installed prior to loading them. Installing them individually is simple using the install.packages("nameofthepackage", dependencies = T) function, which can also be used to update the packages.

```{r setup, message=F, warning=F}

# install.packages("raster") # exemple

# rm(list=ls()) # clean R environment
library(sf) # for spatial simple feature objects
library(raster) # for spatial raster data
library(tidyverse) # for tidy data handling 
library(data.table) # for %like% operator (eventually faster data operations?)
library(kableExtra) # for clean html output tables
```

## 1.2. Inputs and Outputs

Specify directory containing all sub-directories to explore for the original .TIF data from the digitalglobe acquisitions.

```{r finding images}
imgs_dir = "../data/digitalglobe_image/" # directory containing all satellite images (can be in different folders, to be found with recursive search)
```


## 1.3. Define functions

There are no functions to define prior to running this script.


# 2. Data summary

This step is used to catalog all the data from the specified input directory.  
  
Notes:  
March 2021: Only 3 pairs of MUL/PAN images

```{r image search}
# extract date and data type from the filenames
imgs = as.data.frame(
  dir(imgs_dir, recursive = T, pattern = ".TIF$|.tif$")
  ) # creates a dataframe of all image names 

imgs = imgs %>%
  mutate(date = lubridate::ymd_hms(str_match(imgs[,1], ".*(MUL|PAN)/(.*)(-M|-P)2AS")[,3]))
imgs_MUL = imgs %>% 
  filter(str_detect(imgs[,1], "_MUL"))
imgs_PAN = imgs %>% 
  filter(str_detect(imgs[,1], "_PAN"))
```

## 2.1. Multispectral images overview

Let's read the basic metadata from the multispectral (MUL) images. This will be done with a for loop that reads the rasters, retrieve basic metadata and assign them to a tibble object.

```{r MUL formats, results= "hide"}
MUL_summary=tibble() # creates an empty tibble (dataframe)
for(i in 1:length(imgs_MUL[,1])){ # fills the tibble with information from the raster, including resolution and number of layers
  r = brick(paste0(imgs_dir,"/",imgs_MUL[,1][i])) # creates the raster bricks
  r_summary = cbind(print(res(r)[1]),print(nlayers(r)),print(crs(r)),print(imgs_MUL[i,]))
  MUL_summary = rbind(MUL_summary,r_summary)
}

colnames(MUL_summary) = c("res","nlayers","crs","path","date") # changes the variable names

MUL_summary = MUL_summary %>%  # adds information concerning the product ID from it's name.
  mutate(part = str_match(MUL_summary$path, ".*_01_(.*)(.TIF|.tif)")[,2]) %>% 
  mutate(id = str_match(MUL_summary$path, ".*AS(_....-|-)(.*)_P")[,3]) %>% 
  mutate(id_part = paste0(id,"_",part)) %>% 
  as_tibble()
```

The following chunk prints the metadata tibble into a kable output format and writes it's content into a csv file
```{r print MUL formats}
kable(MUL_summary) %>% 
  kable_styling() %>% 
  scroll_box(width="1000px",height="500px")
write.csv(MUL_summary, "../data/multispectral_summary.csv") # writes the table in the project's data folder
```

We seem to have the same CRS information for all archive data, except for the newer acquisitions which are in a longlat projection.  
  
The number of bands and their resolution is also variable from **4 to 8 bands**, going from **1.2 to 2.4 meters** in the archives, and 1.8e-5 degrees in the new acquisitions.  
  
The time of capture is always between 16:00 and 16:40 GMT, which means **11:00 to 11:40 local time**.  
  
The dates range from **2009 to 2019**.  


## 2.2. Panchromatic images overview

The same methods will be applied to panchromatic images in order to obtain the same metadata.

```{r PAN formats, results= "hide"}
PAN_summary=tibble() # creates an empty tibble (dataframe)
for(i in 1:length(imgs_PAN[,1])){ # fills the tibble with information from the raster, including resolution and number of layers
  r = brick(paste0(imgs_dir,"/",imgs_PAN[,1][i])) # creates the raster bricks
  r_summary = cbind(print(res(r)[1]),print(nlayers(r)),print(crs(r)),print(imgs_PAN[i,]))
  PAN_summary = rbind(PAN_summary,r_summary)
}

colnames(PAN_summary) = c("res","nlayers","crs","path","date") # changes the variable names

PAN_summary = PAN_summary %>%  # adds information concerning the product ID from it's name.
  mutate(part = str_match(PAN_summary$path, ".*_01_(.*)(.TIF|.tif)")[,2]) %>% 
  mutate(id = str_match(PAN_summary$path, ".*AS(_....-|-)(.*)_P")[,3]) %>% 
  mutate(id_part = paste0(id,"_",part)) %>% 
  as.tibble()
```

```{r print PAN formats}
kable(PAN_summary) %>% 
  kable_styling() %>% 
  scroll_box(width="1000px",height="500px")
write.csv(PAN_summary, "../data/panchromatic_summary.csv") # writes the table in the data folder
```


## 2.3. Products summary

Let's merge some information from both overviews to make a more synthetic table

```{r}
summary_df = cbind(PAN_summary$id_part,as.character(PAN_summary$date),PAN_summary$res,MUL_summary$res,MUL_summary$nlayers, as.character(MUL_summary$crs))

colnames(summary_df) = c("id","date","pan_res","mul_res","mul_bands","crs")

kable(summary_df) %>% 
  kable_styling() %>% 
  scroll_box(width="1000px",height="500px")
write.csv(summary_df, "../data/images_summary.csv")
```


# 3. Metadata parsing

Let's parse the .IMD files of those images to get additionnal metadata of interest for radiometric calibration and atmospheric correction. Both multispectral and panchromatic images will be treated at the same time.  
  
The following chunk will use a for loop to look for specific strings within the .IMD files of the digitalglobe products. They will then be printed to a kable table and written into a .csv file.

```{r imd parsing}
imds = as.tibble(dir(imgs_dir, recursive = T, pattern = ".IMD$", full.names=T)) # finds all .IMD files
imds = imds %>%  # obtain date and product information
  mutate(date = lubridate::ymd_hms(str_match(imds$value, ".*(MUL|PAN)/(.*)(.-M|-P)2AS")[,3])) %>% 
  mutate(part = str_match(imds$value, ".*_01_(.*).IMD")[,2]) %>% 
  mutate(id = str_match(imds$value, ".*AS-(.*)_P")[,2]) %>% 
  mutate(id_part = paste0(id,"_",part))

radiometric_calibration_parameters = data.frame() # dataframe to fill with calibration parameters

for(i in 1:length(imds$value)){ # for every IMD files
imdx = as.tibble(read.csv(paste0(imds$value[i]), sep="=", head=F)) # read as csv

bands = gsub("BAND_","", droplevels( # find band IDs
  as.factor(subset(imdx, imdx$V1 %like% "BEGIN" & imdx$V2 %like% "BAND" )$V2)
))

abscalfactor = as.numeric( # find their abscalfactor
  gsub(";","", droplevels(
  as.factor(subset(imdx, imdx$V1 %like% "absCalFactor")$V2)
)))

effectivebandwidth = as.numeric( # find their effective bandwidth
  gsub(";","", droplevels(
  as.factor(subset(imdx, imdx$V1 %like% "effectiveBandwidth")$V2)
)))

solar_zenith = 90-as.numeric( # define solar zenith angle based on mean sun elevation angle
  gsub(";","", droplevels(
    as.factor(subset(imdx, imdx$V1 %like% "meanSunEl")$V2)
)))

sats = as.character( # define satellite name
  gsub(";","", droplevels(
    as.factor(subset(imdx, imdx$V1 %like% "satId")$V2)
)))

view_angle = 90-as.numeric( # define satellite viewing angle base on mean satellite elevation angle
  gsub(";","", droplevels(
    as.factor(subset(imdx, imdx$V1 %like% "meanSatEl")$V2)
)))

id_part = imds$id_part[i] # identify the part

tdi = as.factor(
  gsub(";","", droplevels(
    as.factor(subset(imdx, imdx$V1 %like% "TDILevel")$V2)
  )))

# pass_angle = as.numeric(
#   gsub(";","", droplevels(
#     as.factor(subset(imdx, imdx))
#   ))
# )        

imdx_params = tibble(bands,abscalfactor, effectivebandwidth, id_part, solar_zenith, sats, view_angle, tdi) # make a dataframe
imdx_params$date = imds$date[i] # add the date

radiometric_calibration_parameters = rbind(radiometric_calibration_parameters, imdx_params) # add to the table
}
```

```{r write MUL rcp}
kable(radiometric_calibration_parameters) %>% 
  kable_styling() %>% 
  scroll_box(width="1000px",height="500px")

write.csv(radiometric_calibration_parameters, "../data/radiometric_calibration_parameters.csv") # writes the table in the data folder
```


# 4. MUL Visualisation  (data specific CEGRIM)
Let's visualize the data contained inside the raw rasters by visualising raster images.  Archive and new acquisitions will be treated separately.

## 4.1. Quickbird MUL 4b 2.4m

```{r }
r = brick(paste0(imgs_dir,"/",last(imgs_MUL[,1])))
values(r)[values(r) < 1] = NA # make 0s into NA
r = trim(r) # adjusts the extent
names(r) = c("C","B","G","Y","R","RE","N2","N") # name the bands
plot(r) # display the data
```

Values are expressed in *digital numbers (DN)* which do not reflect precise radiometrics measurements. They **must be transformed into radiometric units** by absolute radiometric calibration (step 02).
